Change `rating` to `Float`

+++ src/main/java/ua/mibal/booking/model/dto/HotelDto.java
@@ -30,7 +30,7 @@ public record HotelDto(
         String name,
         List<Photo> photos,
         Location location,
-        Integer rating,
+        Float rating,
         BigDecimal minCost
 ) {
 }
